AttendeeIndex
Created a view for the attendee should they choose to use the web app to queue songs instead of using sms.  Using similar styling as the IndexHost I generated a jumbotron heading that contains a h2 reading Queue a song. The button 'search for a song' populates a modal where a song title can be entered to search for a song.