Attendee Queue
Working on the front end for generating the current queue of songs to the attendees view page. Added a div container called queueListContainer for each queue item. This will include song information, an image, song information, and a playback bar to see the song length and playback, and a like and dislike button for attendees to vote on. Working on general layout of queue items as well as taking away play button and volume control functionality.