Song Search Result
Using a for loop returned the song titles of each of the five search results and put them in a list<string>. Put it into a ViewBag and used  a foreach loop to display each title. Need to create a button for each for users to select the correct song.